fixed the debug code which was broken due to the switch to vector2, cleaned up comments, made filter names match class names, fixed size_t to int warnings

+++ CompSci20_PoolGame/Players.cpp
@@ -54,5 +54,5 @@ int Players::getCurrentIndex() const
 
 int Players::getPlayerCount() const
 {
-	return m_gamePlayers.size();
+	return static_cast<int>(m_gamePlayers.size());
 }
